Move get_move to Player from TicTacToe
TicTacToe is responsible for the high-level gameplay.
