+++ spec/lib/tic_tac_toe_spec.rb
@@ -42,24 +42,14 @@ RSpec.describe 'TicTacToe' do
 
       board = spy('board')
       tic_tac_toe.instance_variable_set(:@board, board)
+      player_one = tic_tac_toe.instance_variable_get(:@player_one)
 
       expect(tic_tac_toe).to receive(:display_move_instruction)
       expect(tic_tac_toe).to receive(:display_board).twice
-      expect(tic_tac_toe).to receive(:get_move).and_return(position_str)
+      expect(player_one).to receive(:get_move).and_return(position_str)
       expect(board).to receive(:make_move).with('X', position_str)
 
       tic_tac_toe.start
-    end
-  end
-
-  context 'when get_move is called' do
-    it 'should call UserInterface#get_user_input and receive the input' do
-      input_text = 'input text'
-      allow_any_instance_of(Kernel).to receive(:gets).and_return(input_text)
-
-      user_move = tic_tac_toe.get_move
-
-      expect(user_move).to eql input_text
     end
   end
 end
